refactor(application-server): use the other factory method
When creating an UserOAuth2Account, The code was trying to construct the
response class using the newly created entity as if the relations where
queried.
Replace the use of the `fromEntity` factory constructor to the
`fromResponseDtos` one.
This allow use the usage of the already queries user and service instead
of needing an extra query to join them to the UserOAuth2Account entity.